associando categorias as movimentacoes ManyToMany

+++ src/main/java/br/com/alura/jpa/modelo/Conta.java
@@ -1,5 +1,7 @@
 package br.com.alura.jpa.modelo;
 
+import java.math.BigDecimal;
+
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
@@ -14,7 +16,7 @@ public class Conta {
     private String titular;
     private Integer agencia;
     private Integer numero;
-    private Double saldo;
+    private BigDecimal saldo;
     
 	public Long getId() {
 		return id;
@@ -40,10 +42,10 @@ public class Conta {
 	public void setNumero(Integer numero) {
 		this.numero = numero;
 	}
-	public Double getSaldo() {
+	public BigDecimal getSaldo() {
 		return saldo;
 	}
-	public void setSaldo(Double saldo) {
+	public void setSaldo(BigDecimal saldo) {
 		this.saldo = saldo;
 	}
